+++ src/main.ts
@@ -23,17 +23,18 @@ export const sql = new Sequelize({
 // 	password: "1234"
 // })
 
-secure: if (getConfig("enable_pop3s", false) || getConfig("enable_smtps", false)) {
+secure: if(getConfig("enable_pop3s", false) || getConfig("enable_smtps", false)) {
 	let tlsKey: Buffer, tlsCert: Buffer
 	try {
 		tlsKey  = await readFile(getConfig("tls_key",  "cert/privkey.pem"))
 		tlsCert = await readFile(getConfig("tls_cert", "cert/fullchain.pem"))
-	} catch (ignore) {
+	} catch(ignore) {
 		break secure
 	}
 
-	if (getConfig("enable_pop3s", false)) new POP3Server(getConfig("pop3s_port", 995), true, tlsKey, tlsCert) // Port 110 for regular POP3, 995 for POP3S
+	if(getConfig("enable_pop3s", false)) new POP3Server(getConfig("pop3s_port", 995), true, tlsKey, tlsCert) // Port 110 for regular POP3, 995 for POP3S
+	if(getConfig("enable_smtps", false)) new SMTPServer(getConfig("smtps_port", 465), true, tlsKey, tlsCert) // Port 25 for regular SMTP, 465 for SMTPS
 }
 
-if (getConfig("enable_smtp", true)) new SMTPServer(getConfig("smtp_port", 25)) // Port 25 for regular SMTP, 465 for SMTPS
-if (getConfig("enable_pop3", true)) new POP3Server(getConfig("pop3_port", 110), false) // Port 110 for regular POP3, 995 for POP3S
+if(getConfig("enable_smtp", true)) new SMTPServer(getConfig("smtp_port", 25), false) // Port 25 for regular SMTP, 465 for SMTPS
+if(getConfig("enable_pop3", true)) new POP3Server(getConfig("pop3_port", 110), false) // Port 110 for regular POP3, 995 for POP3S
